# packages/shared/openapi.json
{
  "components": {
    "schemas": {
      "AccessUpdate": {
        "properties": {
          "allowed_agent_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Agent Ids"
          },
          "allowed_tenant_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Tenant Ids"
          },
          "brand_safe": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Safe"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntryStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "title": "AccessUpdate",
        "type": "object"
      },
...
      "CatalogCreate": {
        "description": "Provider creates a catalog (AC49).",
        "properties": {
          "category": {
            "default": "",
            "maxLength": 120,
            "title": "Category",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/CatalogVisibility",
            "default": "private"
          }
        },
        "required": [
          "name"
        ],
        "title": "CatalogCreate",
        "type": "object"
      },
      "CatalogOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "entry_count": {
            "default": 0,
            "title": "Entry Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "integer"
          },
          "shared_agent_ids": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Shared Agent Ids",
            "type": "array"
          },
          "visibility": {
            "$ref": "#/components/schemas/CatalogVisibility"
          }
        },
        "required": [
          "id",
          "provider_id",
          "name",
          "category",
          "visibility"
        ],
        "title": "CatalogOut",
        "type": "object"
      },
...
      "CatalogType": {
        "enum": [
          "event",
          "place",
          "opportunity",
          "offer",
          "itinerary"
        ],
        "title": "CatalogType",
        "type": "string"
      },
      "CatalogUpdate": {
        "description": "Rename / re-categorise / re-publish a catalog (AC49). Only provided fields change.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "visibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogVisibility"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "CatalogUpdate",
        "type": "object"
      },
...
      "EntryContentUpdate": {
        "description": "Provider-owned content edit (AC29). All fields optional; only provided ones change.",
        "properties": {
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          },
          "custom_sections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CustomSection"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Sections"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "destination": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "highlights": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlights"
          },
          "market_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "EntryContentUpdate",
        "type": "object"
      },
      "EntryCreate": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "catalog_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id"
          },
          "custom_sections": {
            "items": {
              "$ref": "#/components/schemas/CustomSection"
            },
            "title": "Custom Sections",
            "type": "array"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "destination": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Destination",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "market_tags": {
            "items": {
              "type": "string"
            },
            "title": "Market Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/CatalogType"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "required": [
          "type",
          "title",
          "destination"
        ],
        "title": "EntryCreate",
        "type": "object"
      },
      "EntryOut": {
        "properties": {
          "asset_keys": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Asset Keys",
            "type": "array"
          },
          "attributes": {
            "additionalProperties": true,
            "default": {},
            "title": "Attributes",
            "type": "object"
          },
          "brand_safe": {
            "title": "Brand Safe",
            "type": "boolean"
          },
          "catalog_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Id"
          },
          "custom_sections": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CustomSection"
            },
            "title": "Custom Sections",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "display_status": {
            "$ref": "#/components/schemas/DisplayStatus"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "highlights": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "market_tags": {
            "items": {
              "type": "string"
            },
            "title": "Market Tags",
            "type": "array"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "integer"
          },
          "review_reason": {
            "default": "",
            "title": "Review Reason",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/CatalogType"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "description",
          "destination",
          "market_tags",
          "status",
          "brand_safe",
          "provider_id",
          "display_status"
        ],
        "title": "EntryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Walsh Content Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/catalog": {
      "get": {
        "operationId": "list_for_agent_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "destination",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Destination"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CatalogType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntryOut"
                  },
                  "title": "Response List For Agent Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List For Agent",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "create_entry_catalog_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Entry",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/catalog/{entry_id}": {
      "delete": {
        "operationId": "delete_entry_catalog__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Entry",
        "tags": [
          "catalog"
        ]
      },
      "get": {
        "operationId": "get_for_agent_catalog__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get For Agent",
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "operationId": "set_access_catalog__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Access",
        "tags": [
          "catalog"
        ]
      },
      "put": {
        "description": "Edit an entry's structured content (AC29). Provider owns the entry; access/status stay on\nthe PATCH endpoint. Any content change bumps ``content_version`` so copies flag (AC33).",
        "operationId": "update_content_catalog__entry_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryContentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Content",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/catalogs": {
      "get": {
        "operationId": "list_own_catalogs_catalogs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogOut"
                  },
                  "title": "Response List Own Catalogs Catalogs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Own Catalogs",
        "tags": [
          "catalogs"
        ]
      },
      "post": {
        "operationId": "create_catalog_catalogs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Catalog",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/accessible": {
      "get": {
        "description": "Catalogs an agent may browse: every public one + the private ones shared with them (AC49).\nThe single catalog-level gate — entry/item reads then flow through services/visibility.",
        "operationId": "list_accessible_catalogs_catalogs_accessible_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogOut"
                  },
                  "title": "Response List Accessible Catalogs Catalogs Accessible Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Accessible Catalogs",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/{catalog_id}": {
      "delete": {
        "operationId": "delete_catalog_catalogs__catalog_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Catalog",
        "tags": [
          "catalogs"
        ]
      },
      "patch": {
        "operationId": "update_catalog_catalogs__catalog_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Catalog",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/{catalog_id}/entries": {
      "get": {
        "description": "Browse a catalog's entries + their items (AC50). A provider sees their own catalog's\nentries; an agent sees them only if the catalog is accessible (and per-entry expiry/off-limits\nstill apply via the visibility choke-point).",
        "operationId": "list_catalog_entries_catalogs__catalog_id__entries_get",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntryOut"
                  },
                  "title": "Response List Catalog Entries Catalogs  Catalog Id  Entries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Catalog Entries",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/catalogs/{catalog_id}/share": {
      "put": {
        "description": "Set the agents a private catalog is shared with (AC49). Only real tourism agents are kept.",
        "operationId": "share_catalog_catalogs__catalog_id__share_put",
        "parameters": [
          {
            "in": "path",
            "name": "catalog_id",
            "required": true,
            "schema": {
              "title": "Catalog Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogShareUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Share Catalog",
        "tags": [
          "catalogs"
        ]
      }
    },
...
  }
}